first person implimented

+++ src/math/pythag.h
@@ -2,6 +2,6 @@
 #ifndef _PYTHAG_
 #define _PYTHAG_
 
-float PythagoreanTheorum(float, float, float);
+double PythagoreanTheorum(float, float, float);
 
 #endif /* _PYTHAG_ */
